requirements.txt
Flask==2.3.2
keras==3.3.3
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.0.1
tensorflow==2.12.0
Python==3.11.4
pickles